+++ requirements.txt
@@ -9,4 +9,4 @@ toml
 tqdm
 #numpy
 #protobuf==3.20.2
-#dequantize_nbit_ops @git+https://github.com/wejoncy/Nbit_DQ_GemV.git@0.1#egg=dequantize_nbit_ops
+XbitOps @git+https://github.com/wejoncy/XbitOps.git
